feature(Address): Show formatting address in msgbox
